+++ apps/server/CHANGELOG.md
@@ -7,7 +7,7 @@
 - Updated dependencies [2027d29]
   - @sundaeswap/capacity-exchange-core@1.1.0
   - @sundaeswap/capacity-exchange-providers@1.3.0
-  - @capacity-exchange/midnight-node@1.0.1
+  - @sundaeswap/capacity-exchange-nodejs@1.0.1
 
 ## 1.0.2